Conflicts:
	MakeDefaults/MakeStdLibDefs.x86-linux
	MakeDefaults/MakeStdLibRules.x86-linux
	Makefile.gcc
	Makefile.linux
	Source/Core/BareMetal/L0Types/Makefile.gcc
	Source/Core/BareMetal/L1Portability/Environment/Linux/Makefile.gcc
	Source/Core/BareMetal/L1Portability/FastPollingMutexSem.cpp
	Source/Core/BareMetal/L1Portability/Makefile.gcc
	Source/Core/BareMetal/L2Objects/Makefile.gcc
	Source/Core/BareMetal/L3Streams/IOBuffer.cpp
	Source/Core/BareMetal/L3Streams/Makefile.gcc
	Source/Core/BareMetal/L4Configuration/Makefile.gcc
	Source/Core/BareMetal/L4Messages/Makefile.gcc
	Source/Core/BareMetal/L5GAMs/GAM.h
	Source/Core/BareMetal/L5GAMs/Makefile.gcc
	Source/Core/BareMetal/L5GAMs/RealTimeApplication.cpp
	Source/Core/BareMetal/L5GAMs/RealTimeApplicationConfigurationBuilder.cpp
	Source/Core/BareMetal/Makefile.gcc
	Source/Core/FileSystem/L1Portability/Environment/Linux/Makefile.gcc
	Source/Core/FileSystem/L1Portability/Makefile.gcc
	Source/Core/FileSystem/Makefile.gcc
	Source/Core/Makefile.gcc
	Source/Core/Scheduler/L1Portability/Environment/Linux/Makefile.gcc
	Source/Core/Scheduler/L1Portability/Makefile.gcc
	Source/Core/Scheduler/Makefile.gcc
	Test/Core/BareMetal/L0Types/Makefile.gcc
	Test/Core/BareMetal/L1Portability/Makefile.gcc
	Test/Core/BareMetal/L2Objects/Makefile.gcc
	Test/Core/BareMetal/L2Objects/ReferenceContainerTest.cpp
	Test/Core/BareMetal/L2Objects/ReferenceContainerTest.h
	Test/Core/BareMetal/L3Streams/Makefile.gcc
	Test/Core/BareMetal/L4Configuration/Makefile.gcc
	Test/Core/BareMetal/L4Messages/Makefile.gcc
	Test/Core/BareMetal/L5GAMs/Makefile.gcc
	Test/Core/BareMetal/Makefile.gcc
	Test/Core/FileSystem/L1Portability/Makefile.gcc
	Test/Core/FileSystem/Makefile.gcc
	Test/Core/Scheduler/L1Portability/Makefile.gcc
	Test/Core/Scheduler/L5GAMs/Makefile.gcc
	Test/Core/Scheduler/Makefile.gcc
	Test/GTest/BareMetal/L0Types/Makefile.gcc
	Test/GTest/BareMetal/L1Portability/Makefile.gcc
	Test/GTest/BareMetal/L2Objects/Makefile.gcc
	Test/GTest/BareMetal/L2Objects/ReferenceContainerGTest.cpp
	Test/GTest/BareMetal/L3Streams/Makefile.gcc
	Test/GTest/BareMetal/L4Configuration/Makefile.gcc
	Test/GTest/BareMetal/L4Messages/Makefile.gcc
	Test/GTest/BareMetal/L5GAMs/Makefile.gcc
	Test/GTest/BareMetal/Makefile.gcc
	Test/GTest/FileSystem/L1Portability/Makefile.gcc
	Test/GTest/FileSystem/Makefile.gcc
	Test/GTest/Makefile.gcc
	Test/GTest/Scheduler/L1Portability/Makefile.gcc
	Test/GTest/Scheduler/L5GAMs/Makefile.gcc
	Test/GTest/Scheduler/Makefile.gcc

+++ Source/Core/BareMetal/L1Portability/FastPollingEventSem.cpp
@@ -60,17 +60,17 @@ void FastPollingEventSem::Create(const bool wait) {
 }
 
 ErrorManagement::ErrorType FastPollingEventSem::FastWait(const TimeoutType &timeout,
-                                                         float64 sleepTime) const {
+                                                         float32 sleepTime) const {
     ErrorManagement::ErrorType err = ErrorManagement::NoError;
     uint64 ticksStop = timeout.HighResolutionTimerTicks();
     ticksStop += HighResolutionTimer::Counter();
 
     // sets the default if negative
-    if (sleepTime < 0.0) {
-        sleepTime = 1e-3;
+    if (sleepTime < 0.0F) {
+        sleepTime = 1e-3F;
     }
 
-    bool noSleep = IsEqual(sleepTime, 0.0);
+    bool noSleep = IsEqual(sleepTime, 0.0F);
 
     while (*flag == 0) {
         if (timeout != TTInfiniteWait) {
@@ -96,7 +96,7 @@ void FastPollingEventSem::Reset() {
 }
 
 ErrorManagement::ErrorType FastPollingEventSem::FastResetWait(const TimeoutType &timeout,
-                                                              const float64 &sleepTime) {
+                                                              const float32 &sleepTime) {
     Reset();
     return FastWait(timeout, sleepTime);
 }
